Collect TF functionality in separate module
