+++ terraform/authentication.tf.json.template.py
@@ -3,7 +3,7 @@ import shlex
 from azul import (
     config,
 )
-from azul.deployment import (
+from azul.terraform import (
     emit_tf,
 )
