+++ terraform/dynamo.tf.json.template.py
@@ -1,7 +1,7 @@
 from azul import (
     config,
 )
-from azul.deployment import (
+from azul.terraform import (
     emit_tf,
 )
 from azul.version_service import (
