+++ terraform/lambda_layer.tf.json.template.py
@@ -1,11 +1,11 @@
 from azul import (
     config,
 )
-from azul.deployment import (
-    emit_tf,
-)
 from azul.lambda_layer import (
     DependenciesLayer,
+)
+from azul.terraform import (
+    emit_tf,
 )
 
 layer = DependenciesLayer()
