+++ terraform/providers.tf.json.template.py
@@ -1,5 +1,7 @@
 from azul.deployment import (
     aws,
+)
+from azul.terraform import (
     emit_tf,
 )
