+++ terraform/step_function.tf.json.template.py
@@ -5,6 +5,8 @@ from azul import (
 )
 from azul.deployment import (
     aws,
+)
+from azul.terraform import (
     emit_tf,
 )
